Repository, RepositoryImpl, DTO 수정
