+++ src/main/java/roomescape/dto/TimeResponse.java
@@ -1,4 +1,10 @@
 package roomescape.dto;
 
+import roomescape.domain.Time;
+
 public record TimeResponse(Long id, String time) {
+
+    public static TimeResponse from(Time time) {
+        return new TimeResponse(time.getId(), time.getTime());
+    }
 }
