+++ src/main/java/roomescape/repository/TimeRepository.java
@@ -1,68 +1,16 @@
 package roomescape.repository;
 
-import java.sql.PreparedStatement;
 import java.util.List;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
 import roomescape.domain.Time;
 import roomescape.dto.TimeRequest;
-import roomescape.exception.Time.TimeErrorMessage;
-import roomescape.exception.Time.TimeException;
 
-@Repository
-public class TimeRepository {
+public interface TimeRepository {
 
-    private JdbcTemplate jdbcTemplate;
+    List<Time> findAll();
 
-    public TimeRepository(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
+    Time findById(Long id);
 
-    private final RowMapper<Time> rowMapper = (rs, rowNum) -> {
-        Time time = new Time(
-                rs.getLong("id"),
-                rs.getString("time")
-        );
-        return time;
-    };
+    Time create(TimeRequest timeRequest);
 
-    public List<Time> findAll() {
-        String sql = "SELECT id, time FROM time";
-        return jdbcTemplate.query(sql, rowMapper);
-    }
-
-    public Time findById(Long id) {
-        String sql = "SELECT id, time FROM time where id = ?";
-        return jdbcTemplate.queryForObject(sql, rowMapper, id);
-    }
-
-    public Time create(TimeRequest timeRequest) {
-        Time time = new Time(timeRequest.time());
-
-        String sql = "INSERT INTO time(time) VALUES(?)";
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-
-        jdbcTemplate.update(con -> {
-            PreparedStatement ps = con.prepareStatement(
-                    sql,
-                    new String[]{"id"}
-            );
-            ps.setString(1, time.getTime());
-            return ps;
-        }, keyHolder);
-
-        return time.toEntity(keyHolder.getKey().longValue());
-    }
-
-    public int delete(Long id) {
-        String sql = "DELETE FROM time WHERE id = ?";
-        int deleteRows = jdbcTemplate.update(sql, id);
-        if (deleteRows == 0) {
-            throw new TimeException(TimeErrorMessage.NOT_FOUND);
-        }
-        return deleteRows;
-    }
+    int delete(Long id);
 }
